Console and Movement Work
Moved the console stuff around so that scripts will work in and out of
it. Also fixed the triple toggle of ~ for the console window. Also added
noclip and fly commands to the console and they work, though fly has a
tendency to clip going up and down... Also worked a tad bit on the Main
Menu; added picture as a placeholder background.