Jumping Mechanics and Collision Work
Converted everything from the playermovement script into a new addforce
system on the charactermovement script. Everything is much more fluid
now, except for getting stuck on walls (requires better collision
detection). Refined the mesh on the main character to fix the bounding
box issues. Also added a debug option to the console and dev script for
use when printing things to console or other options. Fly and noclip do
not work yet, but the framework is close to implemented. More input
methods were added, though most of them are not used yet.