Added level one, created a pretty background, worked with stuff.
On this day I have added a level one. This was an utter failure as some
extremely stupid part of unity decides that if you zoom in enough to a
2d  camera perspective that you have edited on a 3d zone is now broken.
So now I can't really work with layers as I wanted to as well as
creating levels in unity. I have discovered that in unity you just can
not combine objects into one. This does not mean printing ojects. I am
talking about just straight up combining them to simplify things. But
now unity is a dumb and decided to tell me I couldn't do that. This
means that all level design will have to be done out of unity on some
other program.

Next on the list of things I did was I created a new background. Not
much to say as it was easy as pie and it didn't decide to kill itself
prematurely or later.

I also worked with Jeremys code a little today. However I was very
disappointed to find that is grounded is broken.... so thanks Jeremy for
doing that.

oh and fuck you Cameron.