Main Menu and Options Menu Work
Created an Options menu as a template from the Main menu, functionality
still to come for most things. I also started working on animations for
the text when highlighted. There is an animation controller for each
menu, and then animations for different button states, where highlighted
is the only one that I have changed so far. When highlighted, the text
on the highlighted button increases in size by about 1.1 times. This is
better than increasing the button size because then you can oscillate
between the two stages by hovering your mouse in the middle of the two
sizes. More animation work is necessary because I'm not quite
comfortable with the way everything works so far, so I may tweak the
animation and come up with some new animations for the other states.
Also, there may or may not still be a bug where the first text button on
each menu will glitch out and execute the animation more than once,
specifically when moving your mouse away from hovering over it. In
short, this page of text is just to have a longer commit than Josh,
white still maintaining a level of considerable detail and punctuality
to my previously extremely concise prose for each commit. More to
follow.