Actual Work on the Menu System
Created multiple window managers to handle delegating panel space to
prefabs for both menus and then options panels. Currently only the video
settings is there, but the audio button works as well as video button.
Biggest issue right now is fixing the dropdown menus so that they aren't
dumb-looking; as far as I can tell, it only ignores settings when it is
in a prefab. Also imported and added to the project Unity UI Extensions
to help with the UI design.