If Unity doesn't work, update it!!!!
I updated Unity to version 5.2.2p3 and as a result, had to update some
of the scripts used in the UI extensions to be compatible. Most work was
changing from vertex to mesh. I also made a functional pause menu,
though there isn't much there (not that I have much else to add). Also
did some work with scripts to allow for the pause menu to work fluidly
with the options menu stuff.