Menu and Quest Work
Started outlining the core classes and implementation system for adding
quests. Also started working on the menu system to facilitate the new
quest windows. Made the character stop moving when the window is open,
and added code for the escape key to close windows if they are open.
Will describe the quest system later once things are finalized.