Quest Interaction and Inventory Work
Organized the Inventory system into a two column system for the name on
the left and the quantity on the right. Setup the scripts and example of
a location - based interaction with an object, in this case to start a
quest, though the InteractWith script can be used for anything. Also
fixed a small bug that prevented movement and key presses after
switching levels due to an oversight in the movement prevention system
and load level system. Also added the console to the test quest scene
and the HouseZone scene.