Quest Status in Inventory, GUI Fixes
Setup a massive PITA set of code to display the current quest and its
status in the inventory panel. I'll explain how it works at some point
with the rest of the quest documentation. Also fixed an issue with the
windows not closing/input not responding with opening menus. Turns out I
had the pause listener on a FixedUpdate() which would not sync up frame
per frame with the buttonDown, and there is also an issue with tab as an
inventory key, though I will have to look into that later, as e is not
quite the most optimal button since f has to boe interact.