Setup Lighting and Candle
Moved everything over to dispersed lighting in order to facilitate
dynamic lighting in scenes rather than global illumination. To test, a
spot light was added to the candle on the table and a flicker script was
made to simulate the candle flicker (to some extent). Will work more on
it like the daylight cycle once the main world is created.