Setup Item System
Created Item repository and registry for items with an interfaced class
for each item that can execute functions for each state given. Also
setup the inventory some more to facilitate the new item functionality.
The inventory now refreshes when the OnChanged() delegate is called from
the InventoryData if it is open. Also created a few new prefabs for
generic inventories.