Weapon System and Attack System Created, Some Code Cleanup
Added function to the Iitem Interface to allow for items that do damage.
Created the attack system, most code is built into the playermovement
class (might want to rename it now that it does more stuff). Character
now attacks anything it is facing (if it is in the layer mask) with
different items depending on what is equipped. Also fixed the generic
inventories working and cleaned up some other code.