Inventory Work
Removed the tacky item quantity column from the inventory, have the
quantity display on the left if there is more than 1. Also added a new
Arial font to support unicode characters so that a dot can be shown next
to the equipped item. Also added a function to the InventoryData to
allow other scripts to update the inventory.