Revamped the Quest System
Created a good quest system. This one follows in the footsteps of the
item system that worked out really well. It uses a simple interface for
all quest scripts and then a Quest Dictionary that stores references to
all quest classes. Each quest class can store static variables regarding
the state of the quest, and return from the GetUpdate() function the
string of their status. More documentation may be written soon, though
there is are a lot of comments all ready in the SampleQuest class. Also
moved around some prefabs for organization and tweaked a few other
classes.