Quest Cleanup, More Commands and Functions
Worked out some bugs with the quests not updating in the inventory and
in game when reset; now all quests register their reset function when
being registered to the dictionary. Also setup an ending function that
is called for dialogue systems. Also added a player name that can be
called and used in dialogue systems or anywhere by referencing the
SaveData playerName variable.