Fixed Control Settings, Added Loading Screen and Game Reset Features`
Fixed issues with different keys not loading on the control window. Also
added a loading screen system with a LoadLevel script that gets passed a
name or integer and then a load function so that a level can be loaded
after the loading screen. Also worked on the main menu, and setup the
New Game functionality with the GameReset script to reset data once the
scene is loaded, which required a change to the script execution
sequence to load the GameReset last.