Added Mob AI with the A* Pathfinding Solution
Setup a test scene of the house zone for pathfinding stuff, created an
implementation of the A* path routing system in the 2d environment with
collision detection of everything. Also modified a public script to
serve the needs of the mob AI, currently with the slime texture. There
is still much to be cleaned  up with the movement, and some optimization
could take place, especially with file sizes; much of the scripts
provided with the API are going unused, so they will need to be removed
once they can be determined unnecessary.