Added Linear Dialogue to the System
Setup the methods needed to linear dialogue of paragraphs to be
integrated into the dialogue system using gameobjects now. Also fixed a
bug with destruction of dialogue windows not taking place until after
new object references are setup. For an example, use the Dialogue prefab
in the dialogues folder. The old dialogue system will stay until
everything can be moved off of it, so only the DialogueOption script
should be used now.