Created the Village Scene, Worked on Scene Loading Systems
Created the Village Scene, and got everything set up in the scene for
functional gameplay. In doing so, a prefab was created to hold all of
the objects needed in every scene as a part of the gameplay experience.
Scripts were also created to keep track of position of the character
when it changes scenes due to entering a building, and then return the
player to the position upon entering back into the world. Other small
changes were made to saving and loading scripts; the loading screen is
no longer being used as it wastes extra time loading a scene in-between
scenes.