Setup Variables for Weapons
Implemented weapon items now have damage, range, and speed variables
that are stored and saved. Functionality and usage will come later with
the rest of the attack system (still waiting on textures of items.) Also
added a boarder texture to the minimap to clean things up and make the
minimap more apparent.