Added Fonts, Textures to UI
Imported and setup the PressStart2P font and got most of the prefabs
setup with the new font (only thing missing is the dialogue windows and
menus). The font is a little hard to read and looks cut off at smaller
font sizes, so it might be changed later or fixed somehow. Also fixed a
bug with generic inventories and transferring items to the player's
inventory.