Major Overhaul to the World System
We've revoked the multiple scene system. Everything now happens in one
scene. The transition now takes the character in the z direction to a
different background. Also, everything is now a 3d object, so 3d
colliders are needed for everything. Nearly everything is already
converted to 3D and should work, but more work is to come on other 3d
objects.