Quest Work, Added Carpenter Dialogue
More small code cleanup. Added the Reset() function to the interface for
the quests so that one iteration through the quest dictionary will
execute all reset functions. Also added the carpenter dialogue to the
game and got it setup. Also made some other changes to the mob AI so
that it didn't move when windows are open. Also changed the CheckQuest
to return true if there is nothing to check.