Quest Work Everywhere, Some Attack Work
Worked on the Rebuild Bridge quest. Added dialogue to the carpenter
(should be almost done now). Fixed some state stuff in all quest
scripts. Re-added the function to pass the quest to the savedata so that
the quest status shows up again in the inventory. Also modified the
enemy AI so that it doesn't try to track you when you enter a building
(go up a level). Also added Cooldown times and range calculations to the
attack system and added a cooldown timer on the mobs as well. Also added
a console command to wipe all PlayerPrefs data, causing everything to
pull the default values.