Completely new Item System
The  item system now uses unique object instances of the item class it
is derived from. Each object is stored in lists with dictionaries to
keep track of the item quantity. All functionality from the
itemdictionary is mostly moved into InventoryData. Also added many other
methods to deal with the unique objects and the need to differentiate
between different instances of the same class. Also fixed a bug with
console causing two windows to open up. Also added the overrided
ToString() method to all items to display the item name. weapons also
have an ID with a unique string ID system to allow unique weapons with
customizable data if needed. The Generic Chest has yet to be moved over
to objects, but there is a temporary conversion in place to allow it to
work with the new object system. More documentation will be available on
Trello once the system is done and it's not early in the morning.