Added Item Description Panel to Inventory and Items
Created a panel in the inventory to display information about the
selected item. Items display their name and description. Equippable
items show a button to equip them. Weapons display weapon stats as well.
All methods were setup for the extra information needed. There is now a
function that calls the DisplayInformation() function on items to get
the string array of information and delineates it into the readable
format. All items should follow the format for display information.
Refer to any item class for the new code and its implementation. Also
may have fixed a bug with the input system sometimes freaking out and
losing key strokes.