Moved Generic Inventory to Object-Based Systems
Used the same system from the inventory to create the genericInventory
item window. Also coded a CreateItemObject script to create the item
objects and add them to the generic Inventory with information set.
There is still a bug with the inventory handling unique items and both
inventories not saving correctly. That will be fixed in a later commit.