Implemented Eat and Equip Buttons to Inventory
Added an eat button to the inventory to show up if the item is of type
IFood. Added scripts to each button to execute their respective function
on the item which was passed to them when the item was selected. In
turn, the InformationBox now keeps track of the item and passes it to
the appropriate button when called. The inventory item selection is
still a bit janky, but that may have some fixes if deemed necessary.