Added Window Brightness/Color Script and Prefab
Created a script to change the color of a blue window background based
on the time of day outside. The code is the same as the sun script, so
hopefully everything should line up. One issue that was noticed is that
jumping forward in time does not translate to the correct brightness as
the brightness is determined by a changing slope, not absolute value
dependent on time.