Added Auto-Typing Text to Dialogue
Dialogue text now fills in each character on a timed clock so that it
looks like it is being typed. Also worked on the dialogue window to
pretty it up and make the text look better. Also changed the formatting
on the options menu to scale the button height with the prompted text.