Refactored the Save System
All data is now stored in a JSON formatted file. Serialization and
deserialization take place in the savedata file, with variables stored
in a save object on the JsonFile script. The save object is written to
the SaveData.json file. Inventory data and options are stored on the
file, however inventory and quest data has yet to be moved over.