Shop System Added
Woot. Finally added the shop system to the game. A cost integer was
added to the item interface, and the GenericChest and new GenericShop
were moved into an Inventory interface. GenericShop works halfway
between both GenericChest and InventoryData/Manager. Where the inventory
screen was used and information displayed on the item. Buy button was
added to the bottom where the equip and use button and code was added to
only interface with that button if the shop is open, otherwise the code
ignores it. Much code was reused, so checks were added to determine
whether a shop is open, or just the standard inventory. Also added
storage of shop information to the save json, formatted the same as the
generic inventory data. Items have a cost and is displayed in the shop
window, and players have a gold variable stored now, but there is yet to
be a display of such value in the game, and commands to go with it.