Sprite Work, Scene Work, Background Work
Lots of minor tweaks to improve life around the editor; added a script
to automatically enable all the houses and backgrounds upon runtime in
case an editor forgot to re-enable them. Also setup the sprite sheets
for the new characters added by Cameron yesterday and implemented the
forester texture in the scene. Also moved the textures into their own
Characters folder now. Also setup the cider bar texture and transition
planes correctly, though colliders still need to be added to the room.
Also added a small amount of documentation to the navigation class.